requirements.txt
google-genai
google-generativeai
chromadb
pypdf
python-dotenv
sentence-transformers